breakout CollectionDialog
